requirements.txt
functions-framework==3.*
google-cloud-logging==3.*
requests==2.*
beautifulsoup4==4.*
pymongo==4.*
certifi